backend/LLM/wiki/UpdateWiki/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blockbuster==1.5.26
brotli==1.2.0
certifi==2026.4.22
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==44.0.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.1
fonttools==4.61.1
forbiddenfruit==0.1.4
frozenlist==1.8.0
fsspec==2025.12.0
googleapis-common-protos==1.72.0
gradio_client==0.7.0
graphifyy==0.4.23
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
grpcio-tools==1.75.1
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonschema_rs==0.29.1
kiwisolver==1.4.9
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.0
langchain-deepseek==1.0.1
langchain-huggingface==1.2.2
langchain-openai==1.1.0
langchain-text-splitters==1.1.2
langgraph==1.1.9
langgraph-api==0.5.32
langgraph-checkpoint==3.0.1
langgraph-cli==0.4.7
langgraph-prebuilt==1.0.10
langgraph-runtime-inmem==0.19.1
langgraph-sdk==0.3.13
langsmith==0.4.56
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
networkx==3.6.1
nltk==3.9.4
numpy==1.26.4
openai==2.9.0
openpyxl==3.1.5
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.8
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pillow==10.4.0
propcache==0.4.1
protobuf==6.33.2
pydantic==2.12.5
pydantic-settings==2.14.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlite-vec==0.1.9
sse-starlette==2.1.3
starlette==0.50.0
structlog==25.5.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.12.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-c-sharp==0.23.5
tree-sitter-cpp==0.23.4
tree-sitter-elixir==0.3.5
tree-sitter-go==0.25.0
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-julia==0.23.1
tree-sitter-kotlin==1.1.0
tree-sitter-lua==0.5.0
tree-sitter-objc==3.0.2
tree-sitter-php==0.24.1
tree-sitter-powershell==0.26.3
tree-sitter-python==0.25.0
tree-sitter-ruby==0.23.1
tree-sitter-rust==0.24.2
tree-sitter-scala==0.26.0
tree-sitter-swift==0.0.1
tree-sitter-typescript==0.23.2
tree-sitter-verilog==1.0.3
tree-sitter-zig==1.1.2
truststore==0.10.4
typer==0.21.0
typer-slim==0.21.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==11.0.3
xlrd==2.0.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0